added Add, and Mul method to Vector class

+++ VerbNurbsSharp/Evaluation/Eval.cs
@@ -13,7 +13,7 @@ namespace VerbNurbsSharp.Evaluation
         /// <param name="u">float parameter</param>
         /// <param name="knots">array of nondecreasing knot values</param>
         /// <returns></returns>
-        public static int KnotSpan(int degree, float u, KnotArray knots)
+        public static int KnotSpan(int degree, double u, KnotArray knots)
         {
             return KnotSpanGivenN(knots.Count - degree - 2, degree, u, knots);
         }
@@ -26,7 +26,7 @@ namespace VerbNurbsSharp.Evaluation
         /// <param name="u">parameter</param>
         /// <param name="knots">array of nondecreasing knot values</param>
         /// <returns>the index of the knot span</returns>
-        public static int KnotSpanGivenN(int n, int degree, float u, KnotArray knots)
+        public static int KnotSpanGivenN(int n, int degree, double u, KnotArray knots)
         {
             if (u > knots[n + 1] - Constants.EPSILON)
                 return n;
